# feat: multiple matching routes for `TrieRouter`