--- a/src/router/trie-router/router.test.ts
+++ b/src/router/trie-router/router.test.ts
@@ -9,11 +9,11 @@ describe('Basic Usage', () => {
   it('get, post hello', async () => {
     let res = router.match('GET', '/hello')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('get hello')
+    expect(res.handlers).toEqual(['get hello'])
 
     res = router.match('POST', '/hello')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('post hello')
+    expect(res.handlers).toEqual(['post hello'])
 
     res = router.match('PUT', '/hello')
     expect(res).toBeNull()
@@ -30,7 +30,7 @@ describe('Complex', () => {
     router.add('GET', '/entry/:id', 'get entry')
     const res = router.match('GET', '/entry/123')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('get entry')
+    expect(res.handlers).toEqual(['get entry'])
     expect(res.params['id']).toBe('123')
   })
 
@@ -38,7 +38,7 @@ describe('Complex', () => {
     router.add('GET', '/wild/*/card', 'get wildcard')
     const res = router.match('GET', '/wild/xxx/card')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('get wildcard')
+    expect(res.handlers).toEqual(['get wildcard'])
   })
 
   it('Default', async () => {
@@ -46,22 +46,64 @@ describe('Complex', () => {
     router.add('GET', '/api/*', 'fallback')
     let res = router.match('GET', '/api/abc')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('get api')
+    expect(res.handlers).toEqual(['fallback', 'get api'])
     res = router.match('GET', '/api/def')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('fallback')
+    expect(res.handlers).toEqual(['fallback'])
   })
 
   it('Regexp', async () => {
     router.add('GET', '/post/:date{[0-9]+}/:title{[a-z]+}', 'get post')
     let res = router.match('GET', '/post/20210101/hello')
     expect(res).not.toBeNull()
-    expect(res.handler).toBe('get post')
+    expect(res.handlers).toEqual(['get post'])
     expect(res.params['date']).toBe('20210101')
     expect(res.params['title']).toBe('hello')
     res = router.match('GET', '/post/onetwothree')
     expect(res).toBeNull()
     res = router.match('GET', '/post/123/123')
     expect(res).toBeNull()
+  })
+})
+
+describe('Multi match', () => {
+  const router = new TrieRouter<string>()
+
+  describe('Blog', () => {
+    router.add('ALL', '*', 'middleware a')
+    router.add('GET', '*', 'middleware b')
+    router.add('GET', '/entry', 'get entries')
+    router.add('POST', '/entry/*', 'middleware c')
+    router.add('POST', '/entry', 'post entry')
+    router.add('GET', '/entry/:id', 'get entry')
+    router.add('GET', '/entry/:id/comment/:comment_id', 'get comment')
+    it('GET /', async () => {
+      const res = router.match('GET', '/')
+      expect(res).not.toBeNull()
+      expect(res.handlers).toEqual(['middleware b', 'middleware a'])
+    })
+    it('GET /entry/123', async () => {
+      const res = router.match('GET', '/entry/123')
+      expect(res).not.toBeNull()
+      expect(res.handlers).toEqual(['middleware b', 'middleware a', 'get entry'])
+      expect(res.params['id']).toBe('123')
+    })
+    it('GET /entry/123/comment/456', async () => {
+      const res = router.match('GET', '/entry/123/comment/456')
+      expect(res).not.toBeNull()
+      expect(res.handlers).toEqual(['middleware b', 'middleware a', 'get comment'])
+      expect(res.params['id']).toBe('123')
+      expect(res.params['comment_id']).toBe('456')
+    })
+    it('POST /entry', async () => {
+      const res = router.match('POST', '/entry')
+      expect(res).not.toBeNull()
+      expect(res.handlers).toEqual(['middleware a', 'middleware c', 'post entry'])
+    })
+    it('DELETE /entry', async () => {
+      const res = router.match('DELETE', '/entry')
+      expect(res).not.toBeNull()
+      expect(res.handlers).toEqual(['middleware a'])
+    })
   })
 })