The current router matches a single route. However, in this PR, the router `TrieRouter` matches multiple routes.

Set up router:

```ts
router.add('ALL', '*', 'middleware a')
router.add('GET', '*', 'middleware b')
router.add('GET', '/entry', 'get entries')
```

Access this path:

```
router.match('/entry')
```

Then matches these:

* `middleware a`
* `middleware b`
* `get entries`

Using this router, we can use a normal handler and middleware handlers as the same handler. 

```ts
app.post('/posts', cors(), (c) => c.text('Created!'))
```

## Performance

The performance has been degraded.

**Current Router  TrieRouter**

```
hono x 883,006 ops/sec ±7.97% (66 runs sampled)
itty-router x 165,710 ops/sec ±4.62% (92 runs sampled)
sunder x 314,507 ops/sec ±3.17% (87 runs sampled)
worktop x 213,234 ops/sec ±3.29% (85 runs sampled)
Fastest is hono
✨  Done in 61.96s.
```

<img width="682" alt="18b2dc5ec2eafe96099fb84baa7d1011" src="https://user-images.githubusercontent.com/10682/166838292-f7b08de1-f355-46e4-9484-fa397c49d9ca.png">

**This Router TrieRouter**

```
hono x 666,840 ops/sec ±8.33% (68 runs sampled)
itty-router x 160,915 ops/sec ±5.07% (86 runs sampled)
sunder x 312,998 ops/sec ±4.79% (88 runs sampled)
worktop x 216,689 ops/sec ±3.44% (86 runs sampled)
Fastest is hono
✨  Done in 54.98s.
```

<img width="615" alt="608d8034b30b64668e0dd01fd8ba282d" src="https://user-images.githubusercontent.com/10682/166838346-2a83bc01-94dd-4762-84fb-7dc22ebc8ef9.png">

##  Advantages

But, this router has the advantages:

1. Routing with this the router doesn't slow down so much if we added middleware handlers compared to using the current routers. Currently, it must have handlers for each middleware. It slow down proportionately.
2. Make routing more simple. We don't have to use `use` for middleware. We can use `get`, `post`, `put`, or `delete` as same as a normal handler. Discuss in #189 

## Concern

It's only `TrieRouter`. `RegExpRouter` is not yet implemeted multiple match routing.